lmdb: Use fewer interfaces -- expose UintValue and UintptrValue directly
